requirements/requirements_light.txt
transformers>=4.26.1
datasets>=2.9.0
torchvision>=0.14.1
tqdm>=4.64.1
sentencepiece>=0.1.97